+++ src/main/java/fuzs/stylisheffects/client/StylishEffectsClient.java
@@ -15,8 +15,7 @@ public class StylishEffectsClient {
     public static void onConstructMod(final FMLConstructModEvent evt) {
         EffectScreenHandler handler = new EffectScreenHandler();
         MinecraftForge.EVENT_BUS.addListener(handler::onPotionShift);
-        MinecraftForge.EVENT_BUS.addListener(handler::onInitGuiPost);
-        MinecraftForge.EVENT_BUS.addListener(handler::onDrawBackground);
+        MinecraftForge.EVENT_BUS.addListener(handler::onGuiOpen);
         MinecraftForge.EVENT_BUS.addListener(handler::onDrawScreenPost);
         MinecraftForge.EVENT_BUS.addListener(handler::onRenderGameOverlayPre);
         MinecraftForge.EVENT_BUS.addListener(handler::onRenderGameOverlayText);