+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v1.1.0-1.16.5] - 2021-12-07
+### Added
+- Added two new config options for excluding a container screen from showing active status effects
+### Changed
+- Changed internal mechanics for how vanilla effect rendering is prevented allowing for better mod compatibility
+### Fixed
+- Fixed compatibility with Curios API and Cosmetic Armor Reworked mods
+
 ## [v1.0.1-1.16.5] - 2021-11-19
 ### Changed
 - The vanilla effect renderer now supports multiple columns